Improved project architecture and extracted login card component
