+++ src/app/components/login-card/login-card.component.html
@@ -1,0 +1,27 @@
+<div class="card login-container">
+        <ngb-alert type="danger" *ngIf="errors.length > 0" (close)="errors=[]">
+                <ul class="alert-list" >
+                    <li *ngFor="let error of errors">{{ error }}</li>
+                </ul>
+            </ngb-alert>
+    <div class="card-body">
+        <h4 class="card-title text-center mt-0 mb-4"><ng-content select="[title]"></ng-content></h4>
+        <ng-content select="[body]"></ng-content>
+
+    </div>
+</div>
+
+<footer id="floating-footer">
+    <div class="row">
+        <div class="col-5">
+            Desenvolvido por uma galera top
+        </div>
+        <div class="col-3">
+        </div>
+        <div class="col-4">
+            <h5>Patrocinadores:</h5>
+            <div>ContaAzul</div>
+            <div>Code:Nation</div>
+        </div>
+    </div>
+</footer>
