+++ src/app/components/login-card/login-card.component.scss
@@ -1,0 +1,36 @@
+.login-container {
+  margin: 0 auto;
+  min-width: 200px;
+  max-width: 400px;
+  border-radius: 4px;
+  background-color: #F1F4F9;
+  margin-top: 10em;
+  margin-bottom: 5em;
+}
+  
+.card-body {
+  margin: 0px 20px 0px 20px;
+}
+
+.alert-list {
+  font-size: 14px;
+  padding: 0 0 0 20px;
+  margin: 0;
+}
+
+.login-links {
+  font-size: 16px;
+  div{
+    min-width: 11em;
+    padding: 0;
+  }
+}
+
+#floating-footer{
+  padding: 1em 2em 1em 4em;
+  margin: 0 auto;
+  margin-top: 2em;
+  width: 40em;
+  border-top: 1px solid #e5e5e5;
+}
+
