+++ src/app/components/login-card/login-card.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit, Input } from '@angular/core';
+
+@Component({
+  selector: 'loginCard',
+  templateUrl: './login-card.component.html',
+  styleUrls: ['./login-card.component.scss']
+})
+
+export class LoginCardComponent implements OnInit {
+
+  @Input() errors: Array<any>;
+
+  ngOnInit() {
+    
+  }
+}
