+++ src/app/pages/change-password/change-password.component.html
@@ -1,36 +1,21 @@
-<div class="card login-container">
-    <div class="card-body">
-        <h4 class="card-title text-center mt-0 mb-4">Recuperação de senha</h4>
+<loginCard [errors]="[]">
+    <div title>Recuperação de senha</div>
+    <div body>
+    <form class="form-signin" [formGroup]="form" (ngSubmit)="onSubmit(form.value)">
 
-        <form class="form-signin" [formGroup]="form" (ngSubmit)="onSubmit(form.value)">
+        <div class="row mt-3 mb-3">
+            <label for="new-password">Nova senha</label>
+            <input type="password" id="new-password" class="login-control" required>
+        </div>
+        <div class="row mt-3 mb-3">
+            <label for="new-password-confirm">Confirmar nova senha</label>
+            <input type="password" id="new-password-confirm" class="login-control" required>
+        </div>
 
-            <div class="row mt-3 mb-3">
-                <label for="new-password">Nova senha</label>
-                <input type="password" id="new-password" class="login-control" required>
-            </div>
-            <div class="row mt-3 mb-3">
-                <label for="new-password-confirm">Confirmar nova senha</label>
-                <input type="password" id="new-password-confirm" class="login-control" required>
-            </div>
-
-            <div class="row mt-5 mb-4">
-                <button class="btn btn-lg btn-primary btn-block" type="submit">Salvar senha</button>
-            </div>
-        </form>
+        <div class="row mt-5 mb-4">
+            <button class="btn btn-lg btn-primary btn-block" type="submit">Salvar senha</button>
+        </div>
+    </form>
     </div>
-</div>
-
-<footer id="floating-footer">
-    <div class="row">
-        <div class="col-5">
-            Desenvolvido por uma galera top
-        </div>
-        <div class="col-3">
-        </div>
-        <div class="col-4">
-            <h5>Patrocinadores:</h5>
-            <div>ContaAzul</div>
-            <div>Code:Nation</div>
-        </div>
-    </div>
-</footer>
+</loginCard>
+
