--- src/app/pages/dash/dash.component.spec.ts
@@ -1,24 +1,0 @@
-
-import { fakeAsync, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DashComponent } from './dash.component';
-
-describe('DashComponent', () => {
-  let component: DashComponent;
-  let fixture: ComponentFixture<DashComponent>;
-
-  beforeEach(fakeAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ DashComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(DashComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should compile', () => {
-    expect(component).toBeTruthy();
-  });
-});
