+++ src/app/pages/request-password/request-password.component.html
@@ -1,38 +1,22 @@
-<div></div>
-<div class="card login-container">
-    <div class="card-body">
-        <h4 class="card-title text-center mt-0 mb-4">Recuperação de senha</h4>
+<loginCard [errors]="[]">
+    <div title>Recuperação de senha</div>
 
-        <form class="form-signin" [formGroup]="form" (ngSubmit)="onSubmit(form.value)">
+    <div body>
+    <form class="form-signin" [formGroup]="form" (ngSubmit)="onSubmit(form.value)">
 
-            <div class="row mt-4">
-                <p class="card-text">Para recuperar sua senha, preencha o formulário abaixo. Após, um email contendo as instruções para recuperação de senha será enviado.
-                </p>
-            </div>
-            
-            <div class="row mt-3 mb-3">
-                <label for="input-email">Email da conta existente</label>
-                <input type="email" id="input-email" class="login-control" required>
-            </div>
+        <div class="row mt-4">
+            <p class="card-text">Para recuperar sua senha, preencha o formulário abaixo. Após, um email contendo as instruções para recuperação de senha será enviado.
+            </p>
+        </div>
+        
+        <div class="row mt-3 mb-3">
+            <label for="input-email">Email da conta existente</label>
+            <input type="email" id="input-email" class="login-control" required>
+        </div>
 
-            <div class="row mt-4 mb-4">
-                <button class="btn btn-lg btn-primary btn-block" type="submit">Recuperar senha</button>
-            </div>
-        </form>
+        <div class="row mt-4 mb-4">
+            <button class="btn btn-lg btn-primary btn-block" type="submit">Recuperar senha</button>
+        </div>
+    </form>
     </div>
-</div>
-
-<footer id="floating-footer">
-    <div class="row">
-        <div class="col-5">
-            Desenvolvido por uma galera top
-        </div>
-        <div class="col-3">
-        </div>
-        <div class="col-4">
-            <h5>Patrocinadores:</h5>
-            <div>ContaAzul</div>
-            <div>Code:Nation</div>
-        </div>
-    </div>
-</footer>
+</loginCard>
