--- src/app/service/login-guard.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { LoginGuard } from './login-guard.service';
-
-describe('LoginGuardService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [LoginGuard]
-    });
-  });
-
-  it('should be created', inject([LoginGuard], (service: LoginGuard) => {
-    expect(service).toBeTruthy();
-  }));
-});
